boxes + chemicals + briquettes + materials + trade.
random edits.
added chemistry to the test tube + random to the boxes.
There are few types of boxes with static loot.
hospital briquette.
changes to the MetroMat.
Crafting hospital briquettes with the help of organs.
Craft grenades.
mosinka melee damage.
a workbench for crafts.
The duration of the torch burning is 5 times. The duration of Gorenje burning is 5 times.
the duffel bag increased the capacity, but increased the slowness.